requirements.txt
amqp==2.4.0
Babel==2.6.0
billiard==3.5.0.5
celery==4.2.1
certifi==2018.10.15
chardet==3.0.4
cycler==0.10.0
Flask==0.10.1
flower==0.9.2
gunicorn==19.9.0
idna==2.7
itsdangerous==1.1.0
Jinja2==2.10
kiwisolver==1.0.1
kombu==4.2.2.post1
MarkupSafe==1.0
numpy==1.15.2
opencv-python==3.4.3.18
Pillow==5.3.0
pkg-resources==0.0.0
pyparsing==2.2.2
python-dateutil==2.7.3
pytz==2018.9
pyzbar==0.1.7
redis==3.0.1
requests==2.19.1
selenium==3.14.1
six==1.11.0
splinter==0.9.0
tornado==5.1.1
urllib3==1.23
vine==1.2.0
Werkzeug==0.14.1
